+++ lib/src/commands/flutter/application_mixin.dart
@@ -15,7 +15,7 @@ mixin ApplicationMixin on EnvironmentMixin {
     final List<Map<String, dynamic>> clients = List<Map<String, dynamic>>.from(data['client']);
 
     final Map<String, dynamic> clientData = clients.firstWhere(
-        (element) => element['client_info']['android_client_info']['package_name'] == baseAppId,
+        (element) => element['client_info']['android_client_info']['package_name'] == androidAppId,
         orElse: () => null);
 
     if (clientData == null) {
@@ -47,7 +47,7 @@ mixin ApplicationMixin on EnvironmentMixin {
 
   Future<AndroidApp> getAndroidApp() async {
     final ListAndroidAppsResponse data = await androidAppsApi.list(firebaseProjectName);
-    final AndroidApp app = data.apps.firstWhere((element) => element.packageName == baseAppId, orElse: () => null);
+    final AndroidApp app = data.apps.firstWhere((element) => element.packageName == androidAppId, orElse: () => null);
     if (app == null) {
       throw ArgumentError(
           'This environment has not been created yet. You need to first call: \nflutter:environment create --env $rawEnv');
@@ -57,7 +57,7 @@ mixin ApplicationMixin on EnvironmentMixin {
 
   Future<IosApp> getIosApp() async {
     final ListIosAppsResponse data = await iosAppsApi.list(firebaseProjectName);
-    final IosApp app = data.apps.firstWhere((element) => element.bundleId == baseAppId, orElse: () => null);
+    final IosApp app = data.apps.firstWhere((element) => element.bundleId == iosAppId, orElse: () => null);
     if (app == null) {
       throw ArgumentError(
           'This environment has not been created yet. You need to first call: \nflutter:environment create --env $rawEnv');